requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
EbookLib
beautifulsoup4
edge-tts
python-multipart